api/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
cohere
sentence-transformers==2.2.2
torch==2.0.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
